move analysis task queue and its execution to csbench_analyze.c from csbench_run.c
